Add support for AirVisual Node
