+++ pyairvisual/api.py
@@ -1,9 +1,11 @@
 """Define an object to interact with the AirVisual data API."""
 from typing import Awaitable, Callable, Union
 
+from .const import NODE_URL_SCAFFOLD
+
 
 class API:
-    """Define the "API" endpoint group."""
+    """Define the "API" object."""
 
     def __init__(self, request: Callable[..., Awaitable[dict]]) -> None:
         """Iniitialize."""
@@ -48,6 +50,10 @@ class API:
             longitude: Union[float, str] = None) -> dict:
         """Return data from nearest station (IP or coordinates)."""
         return await self._nearest('station', latitude, longitude)
+
+    async def node(self, node_id: str) -> dict:
+        """Return data from a node by its ID."""
+        return await self._request('get', node_id, base_url=NODE_URL_SCAFFOLD)
 
     async def ranking(self) -> dict:
         """Return a sorted array of selected major cities in the world."""
