+++ pyairvisual/const.py
@@ -1,0 +1,3 @@
+"""Define package constants."""
+API_URL_SCAFFOLD = 'https://api.airvisual.com/v2'
+NODE_URL_SCAFFOLD = 'https://www.airvisual.com/api/v2/node'
