+++ tox.ini
@@ -28,4 +28,4 @@ whitelist_externals = make
 deps = pipenv
 commands=
     make init
-    pipenv run mypy --ignore-missing-imports pyairvisual
+    make typing
